services/usersService/pyproject.toml
[project]
name = "usersservice"
version = "0.1.0"
description = ""
authors = [
    {name = "Lisboa-dev",email = "matheusllisboas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn[standard] (>=0.44.0,<0.45.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "redis (>=7.4.0,<8.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "orjson (>=3.11.8,<4.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
